feat: Initial setting for landing page
