+++ client/src/components/Landing/LandingSection1.jsx
@@ -1,0 +1,16 @@
+import styled from "styled-components";
+
+const Container = styled.div`
+  width: 100%;
+  height: calc(100vh - 69.28px);
+`;
+
+function LandingSection1() {
+  return (
+    <Container>
+      <div></div>
+    </Container>
+  );
+}
+
+export default LandingSection1;
