+++ client/src/components/Landing/LandingSection2.jsx
@@ -1,0 +1,16 @@
+import styled from "styled-components";
+
+const Container = styled.div`
+  width: 100%;
+  height: calc(100vh - 69.28px);
+`;
+
+function LandingSection2() {
+  return (
+    <Container>
+      <div></div>
+    </Container>
+  );
+}
+
+export default LandingSection2;
